refactor(3ds): acquire native sources without PSP dependencies
